Save ImageFeatures details only if something changed
Before, the function store_to_file was writing the file regardless of whether some new information was computed or not.
Now, it instead only writes the file if some new info were added (or if a force option is added)

+++ functions/features.py
@@ -23,7 +23,8 @@ class ImageFeatures:
     What to do when adding a new feature/internal measure to an image:
     - add it as an attribute, either in the "internal values" or in the
         "model features" section. It must be an Optional[type]
-    - add the getters for the value, that also update the attribute
+    - add the getters for the value, that also update the attribute and
+        set self.__modified to True if the value was changed
     - add the getter to the correct location in the dict in to_JSON
     - add a parser in load_details_from_file
     """
@@ -31,6 +32,9 @@ class ImageFeatures:
     def __init__(self, img: MatLike) -> None:
         # Image
         self.__img = img
+
+        # Modified flag
+        self.__modified: bool = False
 
         # Internal values
         self.__px_width_in_mm: Optional[float] = None
@@ -78,24 +82,34 @@ class ImageFeatures:
 
         return self
 
-    def store_to_file(self, path: str) -> None:
+    def store_to_file(self, path: str, force: bool = False) -> None:
         """
-        Stores all the data to a file, in json format
+        Stores all the data to a file, in json format.
+        If all the values were already loaded from a file (no recomputation),
+        the file will not be written by default.
+        The write can occur also if all the values were loaded, if the force
+        flag is set.
 
         ---------------------------------------------------------------------
         PARAMETERS
         ----------
         - path: the path of the file where to write
+        - force: if set, the file will be written regardless of whether the
+            values were computed or not
         """
 
-        with open(path, "w") as f:
-            f.write(self.to_JSON())
+        result = self.to_JSON()
+
+        if force or self.__modified:
+            with open(path, "w") as f:
+                f.write(result)
 
     def __get_px_width_in_mm(self) -> float:
         if self.__px_width_in_mm:
             return self.__px_width_in_mm
 
         self.__px_width_in_mm = get_px_width_in_mm(self.__img)
+        self.__modified = True
         return self.__px_width_in_mm
 
     def __get_px_height_in_mm(self) -> float:
@@ -103,4 +117,5 @@ class ImageFeatures:
             return self.__px_height_in_mm
 
         self.__px_height_in_mm = get_px_height_in_mm(self.__img)
+        self.__modified = True
         return self.__px_height_in_mm
